Revise gps data

+++ STM32F4/Flight_Controller_550/Core/Inc/PID control.h
@@ -34,11 +34,11 @@ typedef struct _PIDSingle
 	float ki;
 	float kd;
 	
-	float reference;
-	float meas_value;
-	float meas_value_prev;
-	float meas_rate;
-	float meas_rate_prev;
+	double reference;
+	double meas_value;
+	double meas_value_prev;
+	double meas_rate;
+	double meas_rate_prev;
 
 	float error;
 	float error_prev;
@@ -64,9 +64,9 @@ typedef struct PDSingle_GPS
 	float kp;
 	float kd;
 
-	float reference;
-	float meas_value;
-	float meas_value_prev;
+	double reference;
+	double meas_value;
+	double meas_value_prev;
 
 	float total_average;
 	float rotating_mem[40];
@@ -107,7 +107,7 @@ void Single_Alt_PID_Calculation(PIDSingle* axis, float set_point_alt, float baro
 void Single_Altitude_PID_Calculation(PIDSingle* axis, float set_point_altitude, float current_altitude);
 void Double_Altitude_PID_Calculation(PIDDouble* axis, float set_point_antitude, float anltitude);
 void Single_GPS_PD_Calculation(PDSingle_GPS* axis, signed int set_point_gps, signed int gps);
-void Double_GPS_PID_Calculation(PIDDouble* axis, float set_point_gps, float gps);
+void Double_GPS_PID_Calculation(PIDDouble* axis, double set_point_gps, double gps);
 
 #ifdef __cplusplus
 }
